requirements.txt
py-algorand-sdk==1.16.0
python-decouple==3.6
requests==2.28.1
